Fix hamburger menu not opening navbar on mobile
The hamburger menu on other pages apart from the homepage page was not
opening the navigation bar when clicked on mobile devices. This commit
fixes the issue by importing the necessary files to the HTML files for
the pages.

This commit also fixed the issue with the page reloading when an anchor
link is clicked because necessary files have been imported in the other
pages.

Fixes #19, #9